chore(refactoring): Move document tests stuff to test/document
