+++ connector-sdk/test/src/main/java/io/camunda/connector/test/document/TestDocument.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.camunda.connector.test;
+package io.camunda.connector.test.document;
 
 import io.camunda.connector.api.document.Document;
 import io.camunda.connector.api.document.DocumentLinkParameters;
