+++ connectors/aws/aws-bedrock/src/test/java/io/camunda/connector/aws/bedrock/mapper/BedrockContentMapperTest.java
@@ -17,8 +17,8 @@ import io.camunda.connector.api.document.DocumentCreationRequest;
 import io.camunda.connector.api.document.DocumentReference;
 import io.camunda.connector.aws.bedrock.model.BedrockContent;
 import io.camunda.connector.document.jackson.DocumentReferenceModel;
-import io.camunda.connector.test.TestDocument;
-import io.camunda.connector.test.TestDocumentFactory;
+import io.camunda.connector.test.document.TestDocument;
+import io.camunda.connector.test.document.TestDocumentFactory;
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import java.time.Duration;
