+++ connectors/aws/aws-bedrock/src/test/java/io/camunda/connector/aws/bedrock/mapper/DocumentMapperTest.java
@@ -14,7 +14,7 @@ import static org.junit.Assert.assertThrows;
 import io.camunda.connector.api.document.Document;
 import io.camunda.connector.api.document.DocumentReference;
 import io.camunda.connector.document.jackson.DocumentReferenceModel;
-import io.camunda.connector.test.TestDocument;
+import io.camunda.connector.test.document.TestDocument;
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import org.junit.jupiter.api.Test;
