+++ connectors/slack/src/test/java/io/camunda/connector/slack/outbound/model/ChatPostMessageDataTest.java
@@ -33,8 +33,8 @@ import io.camunda.connector.api.document.DocumentMetadata;
 import io.camunda.connector.api.document.DocumentReference;
 import io.camunda.connector.api.error.ConnectorInputException;
 import io.camunda.connector.jackson.ConnectorsObjectMapperSupplier;
-import io.camunda.connector.test.TestDocument;
-import io.camunda.connector.test.TestDocumentMetadata;
+import io.camunda.connector.test.document.TestDocument;
+import io.camunda.connector.test.document.TestDocumentMetadata;
 import java.io.IOException;
 import java.time.OffsetDateTime;
 import java.util.List;
